+++ gitserver/bot-templates/golang/waboorrt/meta.go
@@ -1,7 +1,7 @@
 package waboorrt
 
 type MetaInfo struct {
-	Width  int `json:"w"`
-	Height int `json:"h"`
-	Tick   int `json:"tick"`
+	Width  float64 `json:"w"`
+	Height float64 `json:"h"`
+	Tick   int     `json:"tick"`
 }
